[ Refactor ] 리뷰 적용
