+++ src/assets/scss/layouts/result/ResultBackground.scss
@@ -29,7 +29,7 @@
       background: $color-success;
     }
     &--desc {
-      background: url('../../images/banner3.png') no-repeat center center;
+      background: url('../../../images/banner3.png') no-repeat center center;
       background-size: 90%;
       padding-bottom: 15rem;
     }
@@ -51,7 +51,7 @@
       text-align: center;
       padding-top: 6rem;
       padding-bottom: 6rem;
-      background: url('../../images/banner3.png') no-repeat center 45%;
+      background: url('../../../images/banner3.png') no-repeat center 45%;
       background-size: 90%;
       .main {
         font-size: 2.8rem;
